refactor: delete unnecessary view components
