--- src/view/event/event-form-view/event-form-view.js
@@ -1,10 +1,0 @@
-import AbstractView from '../../../framework/view/abstract-stateful-view';
-
-const createEventFormTemplate = () => '<form class="event event--edit" action="#" method="post"></form>';
-
-export default class EventFormView extends AbstractView {
-  get template() {
-    return createEventFormTemplate();
-  }
-
-}
