--- src/view/event/event-form-view/form-details-view/event-form-details-view.js
@@ -1,9 +1,0 @@
-import AbstractView from '../../../../framework/view/abstract-view';
-
-const createEventFormDetailsTemplate = () => '<section class="event__details"></section>';
-
-export default class EventFormDetailsView extends AbstractView {
-  get template() {
-    return createEventFormDetailsTemplate();
-  }
-}
