+++ src/view/event/event-info-view.js
@@ -1,8 +1,32 @@
-import { EVENT_INFO_FORMAT, RENDER_DATE_FORMAT } from '../../constants';
+import {
+  EVENT_INFO_FORMAT,
+  MAX_SELECT_OFFERS,
+  RENDER_DATE_FORMAT,
+} from '../../constants';
 import AbstractView from '../../framework/view/abstract-stateful-view';
 import { calculateDuration, humanizeDate } from '../../utils.js';
 
-const createEventInfoTemplate = ({ eventType, eventCityName, eventStartDate, eventEndDate, eventPrice, isFavorite }) => `<div class="event">
+const createOffer = ({ title, price }) => `<li class="event__offer">
+      <span class="event__offer-title">${title}</span>
+      &plus;&euro;&nbsp;
+      <span class="event__offer-price">${price}</span>
+    </li>`;
+
+const renderOffers = (offers) => {
+  const offersTemplate = [];
+  for (let i = 0; i < offers.length; i++) {
+    const offer = offers[i];
+    if (i === MAX_SELECT_OFFERS) {
+      break;
+    }
+    offersTemplate.push(createOffer(offer));
+  }
+
+  return offersTemplate.join('\n');
+};
+
+
+const createEventInfoTemplate = ({ eventType, eventCityName, eventStartDate, eventEndDate, eventPrice, isFavorite, offers }) => `<div class="event">
   <time class="event__date" datetime = "${eventStartDate}" > ${humanizeDate(eventStartDate, RENDER_DATE_FORMAT)}</time >
   <div class="event__type">
     <img class="event__type-icon" width="42" height="42" src="img/icons/${eventType}.png" alt="Event type icon">
@@ -21,6 +45,7 @@ const createEventInfoTemplate = ({ eventType, eventCityName, eventStartDate, eve
   </p>
   <h4 class="visually-hidden">Offers:</h4>
   <ul class="event__selected-offers">
+    ${renderOffers(offers)}
   </ul>
   <button class="event__favorite-btn ${isFavorite ? 'event__favorite-btn--active' : ''}" type="button">
     <span class="visually-hidden">Add to favorite</span>
@@ -28,9 +53,7 @@ const createEventInfoTemplate = ({ eventType, eventCityName, eventStartDate, eve
       <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z" />
     </svg>
   </button>
-  <button class="event__rollup-btn" type="button">
-    <span class="visually-hidden">Open event</span>
-  </button>
+  <button class="event__rollup-btn" type="button"><span class="visually-hidden">Open event</span></button>
 </div > `;
 
 
@@ -39,10 +62,10 @@ export default class EventInfoView extends AbstractView {
 
   constructor(eventInfo) {
     super();
-    this.eventInfo = eventInfo;
+    this.#eventInfo = eventInfo;
   }
 
   get template() {
-    return createEventInfoTemplate(this.eventInfo);
+    return createEventInfoTemplate(this.#eventInfo);
   }
 }
